Rename itkParticleShapeStatistics to ParticleShapeStatistics.
